// Ui/src/app/application/common/research/enums/model-type-enum.ts
export enum ModelTypeEnum {
    BA = "BA",
    BB = "BB",
    ER = "ER",
    HMN = "HMN",
    NonRegularHierarchic = "NonRegularHierarchic",
    RegularHierarchic = "RegularHierarchic",
    WS = "WS",
    ConnectedRegularHierarchic = "ConnectedRegularHierarchic",
    ConnectedNonRegularHierarchic = "ConnectedNonRegularHierarchic"
}